+++ src/main/java/org/example/fourchak/domain/coupon/service/UserCouponService.java
@@ -11,9 +11,9 @@ import org.example.fourchak.common.error.ExceptionCode;
 import org.example.fourchak.domain.coupon.dto.response.UserCouponResponse;
 import org.example.fourchak.domain.coupon.entity.Coupon;
 import org.example.fourchak.domain.coupon.entity.UserCoupon;
-import org.example.fourchak.domain.coupon.lock.LockService;
-import org.example.fourchak.domain.coupon.lock.NameLockWithDataSource;
-import org.example.fourchak.domain.coupon.lock.NameLockWithJdbcTemplate;
+import org.example.fourchak.domain.coupon.lock.mysql.NameLockWithDataSource;
+import org.example.fourchak.domain.coupon.lock.mysql.NameLockWithJdbcTemplate;
+import org.example.fourchak.domain.coupon.lock.redis.LockService;
 import org.example.fourchak.domain.coupon.repository.CouponRepository;
 import org.example.fourchak.domain.coupon.repository.UserCouponRepository;
 import org.example.fourchak.domain.user.entity.User;
